+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+  - Edge details - Support nuvlabox-peripherals resource
   - ACL - search users escape chars
 
 ## [2.4.5] - 2019-10-10